+++ frontend/app/(private-layout)/project/layout.tsx
@@ -1,19 +1,17 @@
 // app/(private-layout)/project/layout.tsx
-import Navbar from "@/components/navbar";
 import LeftPanel from "@/components/left-panel";
 
 interface ProjectLayoutProps {
   children: React.ReactNode;
-  params: { id?: string };
 }
 
-export default function ProjectLayout({ children, params }: ProjectLayoutProps) {
+export default function ProjectLayout({ children }: ProjectLayoutProps) {
   return (
     <div>
+      {/* Left panel (sidebar) */}
       <div className="sm:sticky bg-background sm:w-fit w-full sm:top-32 sm:mb-0 mb-4">
         <LeftPanel />
       </div>
-      <Navbar projectId={params.id} />
       {children}
     </div>
   );
